Version 6.0.10
Added the necessary hooks so that the winmanager.vim script could more easily integrate with this script. Now the winmanager.vim script should not have to have it's own version of bufexplorer.vim.
Tried to improve performance.
Remember to look at 'help: add-local-help' on how to add the help file into the vim help system.