Version 7.0.1
Thanks to Erwin Waterlander for finding a couple problems. The first problem allowed a modified buffer to be deleted.  Opps! The second problem occured when several files were opened, BufExplorer was started, the current buffer was deleted using the 'd' option, and then BufExplorer was exited. The deleted buffer was still visible while it is not in the buffers list. Opps again!